+++ repos/exposed/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/exposed/AbstractExposedWriteCRUDRepo.kt
@@ -4,9 +4,9 @@ import dev.inmo.micro_utils.repos.UpdatedValuePair
 import dev.inmo.micro_utils.repos.WriteCRUDRepo
 import kotlinx.coroutines.flow.*
 import org.jetbrains.exposed.sql.*
-import org.jetbrains.exposed.sql.statements.InsertStatement
-import org.jetbrains.exposed.sql.statements.UpdateStatement
+import org.jetbrains.exposed.sql.statements.*
 import org.jetbrains.exposed.sql.transactions.transaction
+import java.util.Objects
 
 abstract class AbstractExposedWriteCRUDRepo<ObjectType, IdType, InputValueType>(
     flowsChannelsSize: Int = 0,
@@ -27,10 +27,31 @@ abstract class AbstractExposedWriteCRUDRepo<ObjectType, IdType, InputValueType>(
 
     protected abstract fun InsertStatement<Number>.asObject(value: InputValueType): ObjectType
 
-    protected abstract fun insert(value: InputValueType, it: InsertStatement<Number>)
-    protected abstract fun update(id: IdType, value: InputValueType, it: UpdateStatement)
+    protected abstract fun update(id: IdType, value: InputValueType, it: UpdateBuilder<Int>)
+    protected abstract fun createAndInsertId(value: InputValueType, it: InsertStatement<Number>): IdType
+
+    protected open fun insert(value: InputValueType, it: InsertStatement<Number>) {
+        val id = createAndInsertId(value, it)
+        update(id, value, it as UpdateBuilder<Int>)
+    }
+    @Deprecated(
+        "Replace its \"it\" parameter type with \"UpdateBuilder<Int>\" to actualize method signature. Method with current signature will be removed soon and do not recommended to override anymore"
+    )
+    protected open fun update(id: IdType, value: InputValueType, it: UpdateStatement) = update(
+        id,
+        value,
+        it as UpdateBuilder<Int>
+    )
 
     protected open suspend fun onBeforeCreate(value: List<InputValueType>) {}
+
+    /**
+     * Use this method to do the something with [values]. You may change and output values in that list and return
+     * changed list of pairs
+     */
+    protected open suspend fun onAfterCreate(
+        values: List<Pair<InputValueType, ObjectType>>
+    ): List<ObjectType> = values.map { it.second }
     private fun createWithoutNotification(value: InputValueType): ObjectType {
         return transaction(database) {
             insert { insert(value, it) }.asObject(value)
@@ -40,13 +61,18 @@ abstract class AbstractExposedWriteCRUDRepo<ObjectType, IdType, InputValueType>(
     override suspend fun create(values: List<InputValueType>): List<ObjectType> {
         onBeforeCreate(values)
         return transaction(db = database) {
-            values.map { value -> createWithoutNotification(value) }
+            values.map { value -> value to createWithoutNotification(value) }
+        }.let {
+            onAfterCreate(it)
         }.onEach {
             _newObjectsFlow.emit(it)
         }
     }
 
     protected open suspend fun onBeforeUpdate(value: List<UpdatedValuePair<IdType, InputValueType>>) {}
+    protected open suspend fun onAfterUpdate(
+        value: List<UpdatedValuePair<InputValueType, ObjectType>>
+    ): List<ObjectType> = value.map { it.second }
     private fun updateWithoutNotification(id: IdType, value: InputValueType): ObjectType? {
         return transaction(db = database) {
             update(
@@ -54,7 +80,7 @@ abstract class AbstractExposedWriteCRUDRepo<ObjectType, IdType, InputValueType>(
                     selectById(this, id)
                 }
             ) {
-                update(id, value, it)
+                update(id, value, it as UpdateBuilder<Int>)
             }
         }.let {
             if (it > 0) {
@@ -71,7 +97,9 @@ abstract class AbstractExposedWriteCRUDRepo<ObjectType, IdType, InputValueType>(
 
     override suspend fun update(id: IdType, value: InputValueType): ObjectType? {
         onBeforeUpdate(listOf(id to value))
-        return updateWithoutNotification(id, value).also {
+        return updateWithoutNotification(id, value).let {
+            onAfterUpdate(listOf(value to (it ?: return@let emptyList())))
+        }.firstOrNull().also {
             if (it != null) {
                 _updatedObjectsFlow.emit(it)
             }
@@ -81,9 +109,11 @@ abstract class AbstractExposedWriteCRUDRepo<ObjectType, IdType, InputValueType>(
         onBeforeUpdate(values)
         return (
             transaction(db = database) {
-                values.map { (id, value) -> updateWithoutNotification(id, value) }
-            }.filterNotNull()
-        ).onEach {
+                values.mapNotNull { (id, value) -> value to (updateWithoutNotification(id, value) ?: return@mapNotNull null) }
+            }
+        ).let {
+            onAfterUpdate(it)
+        }.onEach {
             _updatedObjectsFlow.emit(it)
         }
     }