+++ repos/exposed/src/jvmMain/kotlin/dev/inmo/micro_utils/repos/exposed/onetomany/ExposedReadKeyValuesRepo.kt
@@ -1,10 +1,8 @@
 package dev.inmo.micro_utils.repos.exposed.onetomany
 
-import dev.inmo.micro_utils.pagination.*
 import dev.inmo.micro_utils.repos.ReadKeyValuesRepo
 import dev.inmo.micro_utils.repos.exposed.*
 import org.jetbrains.exposed.sql.*
-import org.jetbrains.exposed.sql.transactions.transaction
 
 typealias ExposedReadOneToManyKeyValueRepo<Key, Value> = ExposedReadKeyValuesRepo<Key, Value>
 
@@ -13,54 +11,15 @@ open class ExposedReadKeyValuesRepo<Key, Value>(
     keyColumnAllocator: ColumnAllocator<Key>,
     valueColumnAllocator: ColumnAllocator<Value>,
     tableName: String? = null
-) : ReadKeyValuesRepo<Key, Value>, ExposedRepo, Table(tableName ?: "") {
-    val keyColumn: Column<Key> = keyColumnAllocator()
+) : ReadKeyValuesRepo<Key, Value>, ExposedRepo, AbstractExposedReadKeyValuesRepo<Key, Value>(database, tableName) {
+    override val keyColumn: Column<Key> = keyColumnAllocator()
+    override val ResultRow.asKey: Key
+        get() = get(keyColumn)
+    override val selectByValue: SqlExpressionBuilder.(Value) -> Op<Boolean> = { valueColumn.eq(it) }
+    override val ResultRow.asObject: Value
+        get() = get(valueColumn)
+    override val selectById: SqlExpressionBuilder.(Key) -> Op<Boolean> = { keyColumn.eq(it) }
     val valueColumn: Column<Value> = valueColumnAllocator()
 
     init { initTable() }
-
-    override suspend fun count(k: Key): Long = transaction(database) { select { keyColumn.eq(k) }.count() }
-
-    override suspend fun count(): Long = transaction(database) { selectAll().count() }
-
-    override suspend fun get(
-        k: Key,
-        pagination: Pagination,
-        reversed: Boolean
-    ): PaginationResult<Value> = transaction(database) {
-        select { keyColumn.eq(k) }.paginate(pagination, keyColumn, reversed).map { it[valueColumn] }
-    }.createPaginationResult(
-        pagination,
-        count(k)
-    )
-
-    override suspend fun keys(pagination: Pagination, reversed: Boolean): PaginationResult<Key> = transaction(database) {
-        selectAll().paginate(pagination, keyColumn, reversed).map { it[keyColumn] }
-    }.createPaginationResult(
-        pagination,
-        count()
-    )
-
-    override suspend fun keys(
-        v: Value,
-        pagination: Pagination,
-        reversed: Boolean
-    ): PaginationResult<Key> = transaction(database) {
-        select { valueColumn.eq(v) }.let {
-            it.count() to it.paginate(pagination, keyColumn, reversed).map { it[keyColumn] }
-        }
-    }.let { (count, list) ->
-        list.createPaginationResult(
-            pagination,
-            count
-        )
-    }
-
-    override suspend fun contains(k: Key): Boolean = transaction(database) {
-        select { keyColumn.eq(k) }.limit(1).any()
-    }
-
-    override suspend fun contains(k: Key, v: Value): Boolean = transaction(database) {
-        select { keyColumn.eq(k).and(valueColumn.eq(v)) }.limit(1).any()
-    }
 }